refac: refactor DTO package structure to improve organization
